feat: run programs on hosts (#88)
Related to #51

This PR changes the "Send packet" button to instead have a dropdown list
of "programs", each with a different functionality. Pressing the "Start
program" button starts the program. For now we have three choices:
- No program: just stops the program
- Send ICMP packet: sends a single ICMP packet
- Start echo server: sends ICMP echo packets every X seconds


![image](https://github.com/user-attachments/assets/5ace3471-12bd-4ab9-b1a1-7414f9de764b)


This still has a problem when changing layers: the program continues
running, but with wrong object references. It should be fixed in #89