Fix/keep select when dragging (#110)
This pull request includes several changes to improve the drag
functionality and event handling in the `Viewport` class, as well as a
minor update in the `addDevice` function.

Improvements to drag functionality and event handling:

*
[`src/graphics/viewport.ts`](diffhunk://#diff-8a7a1ebd71e21ac26b73f0dc47262437cfaa066959f22a74a3750b705db06bf7R23):
Added a private `isDragging` property to track the drag state.
*
[`src/graphics/viewport.ts`](diffhunk://#diff-8a7a1ebd71e21ac26b73f0dc47262437cfaa066959f22a74a3750b705db06bf7L36-R51):
Updated event listeners to track drag start and end events, and only
deselect elements on genuine clicks, not drags.

Minor update in `addDevice` function:

*
[`src/types/viewportManager.ts`](diffhunk://#diff-bbcd34ccd1276dd5f80f257a4fc4d064e8a767c4b9be262273b89c4354f5df1cL94-R95):
Commented out the `deselectElement` call and added a TODO comment to
reconsider whether the current element should be deselected.