Improve: Make unselected edges transparent (#180)
This PR tries to unclutter the view of the canvas by making some edges
transparent while selecting devices or edges.

- If the user selects a device, all the edges that are not connected to
the device become transparent.

- If the user selects an edge, all the edges become transparent except
for the selected one.

- If no devices or edges are selected, none of the edges are transparent

Maybe we can add a setting to add/remove this functionality, keeping it
on by default. Some users might want the edges to be at full opacity all
the time.

closes #165