Switching tables (#196)
This PR introduces the implementation of switching tables in the network
simulator, enabling more accurate simulation of switch behavior. The
main updates include:

- **Switching Table Creation and Update:** Switches now maintain a
dynamic switching table that stores MAC address to `DeviceId` (for now,
later it should be changed to interfaces) mappings. These tables are
automatically updated upon receiving packets from connected devices.

- **Packet Forwarding Logic:** Switches use the entries in their
switching tables to forward packets to the correct output interface. If
an entry isn’t found, packets are broadcast as per standard switching
protocols, if it found one but it happens that its output interface
(`DeviceId`) is the same to the input interface, the packet is dropped,
and finally, if an entry with different output interface is found, the
packet is forward to that interface.

- **Refactor of NetworkInterface Location:** The NetworkInterface
component has been moved from DataDevices to ViewDevices, this should
help to integrate interfaces logic to the forwarding packet logic.

---------

Co-authored-by: Pedro Gallino <90009211+pgallino@users.noreply.github.com>
Co-authored-by: Tomás Grüner <47506558+MegaRedHand@users.noreply.github.com>