refactor: updateVisibility modifications (#213)
This PR works on #209 to change the way edge visibility is computed. It
also moves the DFS algorithm we use to the `Graph` class.

The way edge visibility is computed in this PR is by checking if each
device of an edge can reach at least a single visible device. However,
when loops exist in the graph, this can lead to false positives, like:


![image](https://github.com/user-attachments/assets/01855189-aec2-4179-bca6-e2d4fb79e166)


![image](https://github.com/user-attachments/assets/fd16879f-0432-420c-9837-2eb4d8430e96)

We fix this by doing it iteratively many times, ignoring edges that were
marked as invisible in previous iterations.