Support ARP protocol (lightly) (#216)
## Description
This pr add support for using ARP protocol in simulatior. Although this
is a more lightfull version of the protocol.
First of all, as LANs are not implemented, there’s no limit for a device
to send an _ARP request_ to, but only switches can propagate ARP
packets. This means that if a host or a router receive an ARP packet
which wasn’t meant for it, it will discard it. Even routers, which in
reality have the ability to, either resolve the address requesting or to
forward the packet to the corresponding device, as long as the device
corresponds to the same LAN.
Also, every device _ARP table_ is authomatically completed as the device
connects to other devices. This will simplify the UX, as user will not
have to see how the protocol works each time a recently device sends a
packet to another. User will be able to see _ARP request_ and _responds_
packets by running the ARP program. ARP table entries can be deleted or
modified manually by the user.

## Implementation details
### ARP table
The tables are automatically populated each time a device is added to
the network. Adding each entry to the new device's table and updating
the corresponding entries in all existing devices would have a time
complexity of O(n). To avoid this cost, the entries in the table (which
is implemented as a `Map` in the code) are managed as follows:

**(a)** If there is an entry with the device's IP address:

- **(i)** If the `mac` field is an empty string, this indicates that the
user manually deleted the entry. In this case, it is treated as if the
entry does not exist in the ARP table, and the IP address cannot be
resolved.
- **(ii)** If the `mac` field is a non-empty string (presumably a MAC
address), the entry was modified at some point—either manually by the
user or via an ARP Request tool. In both cases, this entry is used to
resolve the IP address.

**(b)** If there is no entry with the device's IP address:
This means the user has never modified the IP entry, either manually or
through an ARP Request. We assume the entry exists and use the viewgraph
to locate the device associated with the IP and, in turn, find its MAC
address.

From the user's perspective, the table will appear exactly as it would
in a real-world scenario.

---------

Co-authored-by: Pedro Gallino <90009211+pgallino@users.noreply.github.com>
Co-authored-by: Manuel Pol <Manuel-Pol@users.noreply.github.com>