Fix/ Routing Table and Payload View (#233)
This PR solves some minor visual changes on the simulator.

### Fewer data shown on payload
As it was previously discussed, some information of the packet payloads
reveal to much data of other layers different from the current one. On
the Network layer, ICMP packets that hold a TCP packet on their payload
just show the type of the payload, not the whole information. However,
ICMP packets do show some data because they are only shown up to the
Network layer


![image](https://github.com/user-attachments/assets/6c576794-a3cf-4903-b971-6ef06abb602a)


![image](https://github.com/user-attachments/assets/3327bb34-cff9-41db-b275-2114428fd0ee)

### Different CSS classes for different tables
With the addition of the TCP Flags table, the modifications broke the
styling of the Router Tables. Now there is a general class for tables
and a separate class for each type of table. This allows to change the
table configuration on each type. There is also a new class for the
recently added ARP tables.


![image](https://github.com/user-attachments/assets/bf331d5e-22d2-46ed-bfb3-a2ca02fa094e)


![image](https://github.com/user-attachments/assets/886bf334-f9ab-483d-b39e-d4afdbafe837)


![image](https://github.com/user-attachments/assets/d70811bc-6573-49db-91e5-ee982c2bc68d)