Dynamic ARP table (#232)
This pull request introduces a mechanism to dynamically update the ARP
table in the UI when changes occur in the underlying data model. The key
changes include adding a listener for ARP table updates, implementing a
refresh method in the ARP table UI component, and integrating these
updates into the `DeviceInfo` class.