feat: render invisible devices as draggable circles (#245)
1. Non-visible devices cannot be selected
Devices that are not visible (including drag circles) can no longer be
selected, connected, or deleted.

2. Merged edges cannot be deleted
Edges that are part of a merged group (i.e., connected to other edges)
cannot be deleted. Only individual, unmerged edges can be removed.

3. Drag circles are for reorganization only
The drag circle functionality is intended solely for moving and
reorganizing devices. It is not possible to connect devices using drag
circles.


![image](https://github.com/user-attachments/assets/8e4a7aec-81bf-44ae-aff6-8c72fd0fe671)


![image](https://github.com/user-attachments/assets/2f84e74d-74f2-4ed3-bd5c-d809f528b9a4)


![image](https://github.com/user-attachments/assets/5e1589f1-ed47-4292-bd5c-800deedd4348)

---------

Co-authored-by: Tomás Grüner <47506558+MegaRedHand@users.noreply.github.com>