fix speed multiplier when setting new network and change 'Switching Table' to 'Forwarding Table' (#263)
### Fix of the bug reported in issue #250 
The bug was fixed by changing the method `GlobalContext.load()`. Now, in
the `GlobalContext.load()` function, if a new `speedMultiplier` is not
provided, the previous value is kept, ensuring the pause/resume state
remains consistent.
### Rename of the "Switching Table" to "Forwarding Table"
Renamed "Switching Table" to "Forwarding Table" to align with
terminology used in official documentation and texts.

Closes #250

Co-authored-by: Pedro Gallino <90009211+pgallino@users.noreply.github.com>