feat: support sending packets from misconfigured router (#271)
This PR allows sending packets from a misconfigured router, enabling us
to show the TTL making routers discard packets.

When a router is misconfigured, the table entry matching with the packet
doesn't lead to the destination. In those cases, routers now try to
resolve the next hop from the interface configured on the entry, instead
of dropping the packet. Packets will still be dropped if the next hop is
not a `NetworkDevice`.

---------

Co-authored-by: Pedro Gallino <90009211+pgallino@users.noreply.github.com>